android/integration_tests/requirements.txt
docker-compose
nose
pyhamcrest
requests
